Refactors foo_snarl.cpp Moves more declarations into headers. Breaks out menu and play_callback. Fixes menu event. Implements Preferencesv2.
